+++ NEWS.md
@@ -1,4 +1,6 @@
 
+# renv (development version)
+
 # renv 1.2.3
 
 * `renv::record()` now enriches each resolved record with the same